01_English_Level_translator/requirements.txt
fastapi
python-multipart
uvicorn
openai==0.28
pydantic
python-dotenv